About Us: Refactor Team section into new component
